feat: xud master password
This is a major commit that introduces new functionality to encrypt the
xud node key with a password which can also be used to encrypt lnd and
(with a separate PR) Raiden.

A new config option `noencrypt` is added and for now defaults to `true`,
meaning that this functionality is off by default. This is to minimize
disruption with existing applications of xud. When `noencrypt` is true,
starting up xud occurs as it does before this PR.

When `noencrypt` is false, we first check if there is an existing node
key file saved to disk. If one exists, we wait to receive a new
`UnlockNode` call to provide us with a password. We use this password to
decrypt the node key file and to call `UnlockWallet` on each configured
lnd instance that is waiting to be unlocked.

If none is found, we wait to receive a new `CreateNode` call to provide
us with a password. We then attempt to call `GenSeed` on a configured
lnd instance to generate an aezeed 24 word mnemonic. We use this
mnemonic to generate our node key, and then encrypt that with our
password and save to disk. We then call `InitWallet` on each configured
lnd instance that is waiting to be unlocked, using the same mnemonic and
password from earlier steps.

As part of this PR, we introduce a new grpc `XudInitService` for calls
when xud is in a state of waiting for a password before it can complete
initialization. New `create` and `unlock` commands are added for
`xucli` as well.

This is a major step towards #912.