feat(raiden): check direct channel before swap
This adds logic to check that we have a direct channel with a peer with
sufficient balance for raiden currencies before accepting or proposing a
swap. Currently multi-hop Raiden payments are not well supported, and
failed payments are not handled well in Raiden. This will prevent any
attempts to swap Raiden tokens without a direct channel.

Closes #1027.