fix: swap sent/received amounts in satoshis
This scales the `amountReceived`, `amountSent`, `makerAmount`, and
`takerAmount` fields to always be denominated in satoshis, even for
currencies where the smallest unit is smaller than  a satoshi. In their
place this adds `takerUnits` and `makerUnits` to the `SwapDeal` type to
track the amounts in each currency's base unit for sending & receiving.

This prevents errors on the rpc layer, where amounts in wei can easily
exceed the limits of `int64`. It also prevents issues in storing these
values in the database, where these fields are likewise represented as
64 bit integers.

This also renames several enum values to use the word "Payment"
instead of "Amount" (as in "PaymentReceived") to avoid any possible
confusion since "amount" is typically used to refer to a number.

Fixes #1063.