feat(swap): persist active swap deals to db
This commit persists the state of swap deals to the database upon each
phase change once a deal has been accepted to. This is done to be able
to prevent loss of funds and recover gracefully from any active swaps
should `xud` crash unexpectedly.

This is the first step towards closing #1079.