refactor(lnd): wait for grpc when verifying conn
This refactors the `verifyConnection` logic to use the grpc
`waitForReady` function to wait for the client to be ready before
attempting a `GetInfo` call to lnd. This cuts down on the number of
calls we need to make to an offline lnd instance to determine when it
has come back online.