feat(swaps): configurable lock buffer time
This replaces the `cltvdelta` config values with a `lockbuffer` option
that specifies the time in hours that we add to the total delay of
the second leg of swaps when determining the minimum final lock time
(final cltv delta in lnd) of the first leg.

It also uses a fixed final lock time of 400 minutes for the final
hop of the second leg of swaps to match the 40 block default in
bitcoin lnd.