feat(raiden): check cltv delay  (#1082)
* feat(raiden): check cltv delay

This commit refines the logic to determine whether a Raiden payment
on the first leg of a swap (taker to maker) extends sufficient cltv
delay according to the maker's specified minimum cltv delay.

This also checks the `token` field of a Raiden resolve request to
ensure that it matches the expected token address for the deal.
It also improves calculation of the `cltvDeltaFactor` to use hardcoded
values for the average time per block for each chain, rather than using
the configured cltv delays for each currency. Changing the configured
values from the defaults could have resulted in a skewed factor.

The cltv delay for the final hop of a raiden payment is set with the
`lock_timeout` parameter on the token payment request. The lock of
an incoming raiden payment is determined by the `expiration` value of
the resolve request. Currently, a warning is logged rather than
rejecting the deal if the `expiration` is insufficient.

Closes #1052.

* feat(swaps): configurable lock buffer time

This replaces the `cltvdelta` config values with a `lockbuffer` option
that specifies the time in hours that we add to the total delay of
the second leg of swaps when determining the minimum final lock time
(final cltv delta in lnd) of the first leg.

It also uses a fixed final lock time of 400 minutes for the final
hop of the second leg of swaps to match the 40 block default in
bitcoin lnd.