feat(swapclient): add timeout to initialization
This adds a timeout to the initialization procedure for swap clients.
Previously, initialization would wait for as long as it took to receive
a response from the client. In cases where raiden or lnd hangs on the
response, xud would hang as well.

Now, xud will wait a fixed time limit before proceeding with
initialization, leaving the client in a not connected state. The swap
client will be marked as connected if and when we get a successful
response.