fix(lnd): waitForReady infinite deadline
This fixes a bug where the first attempt to verify a connection to
lnd would fail on the first attempt due to the `deadline` of 0 for the
`waitForReady` being exceeded immediately with the latest version of
lnd. Subsequent attempts would not exceed the deadline.

Instead, this always uses a deadline of infinity meaning xud will
always wait for lnd to be online before it attempts a `GetInfo` call
to verify the connection. This leverages the recently added
`verifyConnectionWithTimeout` logic to timeout the initial connection
verification if lnd remains offline so that it doesn't hold up the
xud initialization procedure.