fix(swaps): set cltvLimit correctly
This fixes a bug whereby the `makerCltvDelta` was erroneously used for
the `cltvLimit` for the payment from maker to taker. `makerCltvDelta` is
irrelevent to the payment to the taker. Instead, this value is set to
the time lock of the route found from maker to taker for lnd, raiden
currently has no equivalent to `cltvLimit`.

In tests, payments would still fail occasionally due to no route found.
Adding 3 blocks to the `cltvLimit` value consistently resolved these
failures - any fewer than 3 and the payments would still fail. More
investigation is warranted into why this is necessary, it's possible
that there is a bug in the `cltvLimit` implementation within lnd.

Fixes #1158.