feat(swaps): improved remote error handling
This improves the way swap errors are communicated between participants
of a swap. Previously, any error during an attempt to send payment was
communicated to the peer via a `SwapFailedPacket`. However, in cases
where the maker to taker (2nd leg) payment fails, the maker will reject
the 1st leg payment because they were unable to acquire the preimage.
This results in two near-simultaneous payment errors on both peers,
who send `SwapFailedPacket`s to each other for already failed deals
which results in an undesirable `received swap failed packet for
unknown deal` log messages.

This change detects an lnd payment that failed due to being rejected
(`UnknownPaymentHash` error message) and skips sending the
`SwapFailedPacket`. Instead it fails the swap due to a new
`RemoteError` swap failure reason, and updates the error details of
the failed swap if it receives a `SwapFailedPacket` for swaps with
this failure reason.