feat(swaps): dynamic lock buffer
This modifies the logic around calculating the `makerCltvDelta` value
for swaps which specifies the minimum expected lock duration for the
final hop of the first leg. We use the poisson quantile function to
determine a very high probability that the first leg final lock won't
expire before the second leg payment.

This also removes the recently added `lockBuffer` config option in
favor of a `cltvDelta` for lnd that specifies the lock delta that is
used for the final hop of the second leg and gets added to the lock
buffer to determine the final lock delta for the first leg.

Closes #1164.