fix(lnd): wait for macaroon before wallet created
This fixes the logic around initializing `LndClient` for a new lnd
instance (before wallet creation) and new xud instance (before node key
creation). The lnd macaroon file is not created until after its wallet
is created, so we relax the check that disables the client if a macaroon
is not found during initialization. Instead, we mark the client as
`WaitingUnlock` and wait for a successful xud `CreateNode` call to
initialize the lnd wallet. After we've successfully loaded the admin
macaroon, we then continue with attempting to verify the connection.