test: instability simulation tests
This adds a new suite of simulation tests to test how `xud` responds to
instability. It simulates crashes immediately after the maker sends
payment as part of a swap. When the maker comes back online, it will
fetch the preimage for the successful payment and use it to settle
the incoming payment from the interrupted swap.

This tests two key cases, one where the maker's payment goes through
while the maker is offline, and another where the payment goes through
only after the maker has come back online. In the latter case, the maker
must detect that the payment is in a pending state and monitor it in
case it goes through.

Closes #1183.