feat(swapclient): update capacity on channelbalance
This modifies the logic around updating the maximum outbound capacity
for a swap client to occur every time the channel balance is fetched,
regardless of whether `channelbalance` was called by the timer to
periodically update capacity or via an `xud` gRPC request. This will
give us slightly more up-to-date cached balance values without
needing to make any additional calls to the swap clients.