feat(swapclient): heartbeat
This commit shortens the interval for the outbound capacity check timer
from 60 to 3 seconds and sets the client as disconnected any time a
call fails due to an unreachable server. This makes the capacity checks
act like a heartbeat, checking that the server is reachable every few
seconds even in the absence of any other activity.

Previously, we had used a dummy server -> client streaming call and
listened for the `error` event on the lnd side, however with newer
versions of lnd and grpc this is no longer a reliable way to tell when
lnd has gone down.

This also resolves an issue where lnd would get stuck in the
`WaitingUnlock` state if it is stopped while xud is running and comes
back online in the locked state.

Closes #1090.