fix(seedutil): separate aezeed & encryption pwords
This modifies the seedutil tool so that it uses separate passwords for
the aezeed and for encrypting the private key in the keystore.
Previously, a single password was used for both, which is incompatible
with how xud uses them separately with a master password for decryption
and currently always using the default "aezeed" password for the aezeed
itself.

The encryption password, keystore path, and aezeed passwords are now
treated as optional flags.