feat: create eth keystore with master seed
This commit uses the seed generated during the `CreateNode` call to
create the raiden/ethereuem keystore along with the lnd wallets. This
is the next step towards having a single recovery seed and a single
encryption password for all lnd and raiden wallets. The keystore is
encrypted with an empty password instead of the master password until
raiden adds a way to decrypt the keystore without needing to store
the password in a file in plaintext.

A new configuration option `raiden.keystorepath` specifies where xud
should create the keystore directory.

Closes #1242.