feat(cli): wait for client on create/unlock
This adds logic to wait for the xud client to be ready before attempting
the `CreateNode` or `UnlockNode` calls. These calls may be made shortly
after starting xud - for example as part of a bash script - and this
wait of up to 3 seconds may prevent transient failures or race
conditions.