feat(cli): wait for tls cert on create
This adds logic to wait up to 5 seconds for the tls cert file to be
created on the `create` command. Since this command may be called
immediately after xud is launched for the first time and the cert won't
be created until xud is done initializing, this will help avoid
unnecessary failures and repeated attempts.