feat(swaps): raiden claims recovered swap payments
This adds logic to track recovered swap preimages for claiming raiden
payments. A proposed change to raiden has it continuously attempt to
resolve a hash until it gets a response - if xud crashes during or
before the first attempt then raiden will continue trying until xud
comes back online. When it does, xud will check if it has recovered the
preimage and return it to raiden if so, effectively claiming the payment
and preventing loss of funds.

Closes #1251.