feat(resolver): retry flag in error message
This adds a `retry` flag to the resolver error message to indicate
whether the resolver request should be retried. This is set in cases
where xud is not certain that the swap will never be resolved, such as
in cases where a call to send payment is interrupted or does not respond
normally. It also sets a 503 status code when a retry is indicated.

Closes #1282.