fix(p2p): remove error event from Peer
This removes the `error` event from `Peer`. Previously it was only used
to log the error from the listener in `Pool`, however this was
unnecessary and instead the logging can be done directly from `Peer`.

This simplifies the code and also resolves a mysterious crash condition
where an unhandled error from `Peer` would crash xud after a failed
inbound connection via tor.

Closes #1129.