feat(rpc): notify when xud is starting
This changes the xud initialization procedure to start the grpc server
immediately and have it return an `UNAVAILABLE` error with the message
'xud is starting' until all components have finished initializing and
xud is ready to handle rpc calls. It also enhances the xucli output
somewhat in cases where xud is locked or starting.

Closes #1258.