test: pool suite cleanup
This makes some minor cleanup and fixes to the Pool jest test suite.
It also splits the `it connects exactly once if two peers attempt
connections to each other simultaneously` case into two, one where the
first connection is completed before the second connection is attempted,
and another where the connections begin simultaneously. The latter
results in a race condition that occasionally causes the test case to
fail, which will be fixed in a separate commit.