feat(rpc): unlock returns list of locked wallets
This adds a response parameter for the `UnlockNode` call to include a
list of enabled wallets that could not be unlocked. It also uses the
existing `waitForLnd` method to wait briefly for lnd clients to come
online which will help ensure lnd gets unlocked in cases where it is
started near simultaneously with xud. Finally it enhances the `xucli
unlock` output.

Closes #1319.