feat(rpc): add owner parameter for listorders
This commit adds an optional `owner` parameter for the `listorders` command while removing `include_own_orders_parmeter` and replacing its functionality.
The default value for the new parameter is `Both`, other options are `Own` and `Peer`.

The commit also fixes a bug that caused a situation where sell orders were displayed instead of buy orders when the limit was set.

Closes #1323