feat(rpc): RestoreNode call to restore from seed
This adds a new `RestoreNode` call that accepts a 24 word seed mnemonic
to restore an xud key, lnd wallets, and a raiden keystore from the seed.
This call can be extended to also restore lnd channel backups and xud
db backups with additional parameters.

The seedutil tool is extended to support an `encipher` subcommand that
converts a mnemonic to the enciphered seed bytes, which is necessary
for deriving an xud key from aezeed.

Closes #1017. Closes #1020.