feat(config): add maxlimits options
This adds a new `maxlimits` config option that overrides the hardcoded
mainnet limits when set to true. In place of the mainnet limits, the
maximum channel size for a currency's respective network is used as
the maximum order size. These higher limits are now also enforced on
non-mainnet networks regardless of the `maxlimits` value.

Closes #1304.