feat: use deciphered seed for xud key
This changes the way the xud node key is derived from the recovery
mnemonic to use the deciphered seed - which does not contain salt and
certain error checking bytes - rather than the enciphered seed. This
is to ensure that the same node key is derived from a recovery seed
regardless of how it is enciphered using salt and an aezeed password.