feat: restore from backup
This adds an rpm and cli call to allow restoring an xud instance from
backups. This call restores the xud database and pubkey as well as the
lnd wallets and eth keystore. Furthermore it restores from the lnd
static channel backup file which attempts to reconnect to peers with
which we had channels and force close channels, recovering our balance
from the channel.

Co-authored-by: Daniel McNally <mcnallydp@gmail.com>
Co-authored-by: Karl Ranna <karl@karlranna.com>