refactor: asynchronous backup (#1364)
* refactor(backup): listen to lnd connected event

This changes the way the backup process is notified when an lnd client
becomes connected again from checking on an interval to listening
to the `connectionVerified` event which provides immediate notification.

* refactor(backup): async subscriptions and watchers

This makes the backup process begin listening or watching for channel
and database changes asynchronously, instead of one at a time. It also
handles the case where a database file does not yet exist at the time
the backup process is started by waiting for it to be created.