fix(rpc/lnd): lnd already unlocked
This fixes the clients that get added to the `lockedLnd` list in the
`UnlockNode` response. Previously, lnd clients that were already
unlocked were added to this list. This makes it so those are not added,
instead only clients where the unlock call fails or those where the
call can't be made are added to the list.