refactor(nodelist): separation from pool
This refactors the NodeList and Pool classes to have a cleaner
separation of interests between the two classes. The Pool class no
longer touches the NodeInstance class which represents nodes in the
database.

Also, this improves certain operations by reading nodes from
the node list in memory rather than from the database.