refactor: set swap client disconnected on timeout
This explicitly sets the status of a swap client to `Disconnected` after
time runs out for the initial connection verification of a swap client.
Since we wait to make a call until a swap client is "ready" - which
means the grpc client `waitForReady` method finishes in the case of lnd
- we won't experience a failed call right away in cases where lnd is
simply offline. This means the status would remain at `Initialized`
when `Disconnected` is more accurate - at that point we should treat
lnd as if it is offline.