feat(cli): better messages when unlock/create fail
This displays better error messages when `xucli unlock`, `xucli create`,
or `xucli restore` calls fail due to an xud node not existing or
already existing. Instead of the generic `call is not supported by the
current state of xud` message being printed, a more helpful message
is printed informing the user of the state of xud and which call should
be tried instead.

This also returns an error when a wallet that is already unlocked is
attempted to be unlocked again.

Closes #1370.