chore: run 'npm audit fix'
This fixes known bugs and vulnerabilities in dependencies via the `npm
audit fix` command.

`npm audit` still reports one issue with the sequelize package. However,
fixing this requires making an upgrade with breaking changes and the
vulnerability is not a cause for concern at this time as xud controls
all queries made to the database and none are affected by the issue.