refactor: organize/clarify canRouteToPeer logic
This commit attempts to better organize and clarify the `hasRouteToPeer`
logic used to determine whether any possible route exists to pay a peer
for a given currency before enabling trading pairs involving that
currency for that peer.

Previously, lnd and raiden specific logic was located inside the
SwapClientManager class. This has been refactored into the respective
LndClient and RaidenClient classes.

The naming of the method is also changed from "hasRoute" to "canRoute"
to indicate that it only tests whether such a route is possible, not
whether one necessarily exists or is functional. Methods have been
commented to better indicate what they do.