refactor(swapclient): status and initialization
This commit makes several related changes to the `SwapClient` class,
especially in regards to managing the status of a client, with the goal
of making future changes and adding new `SwapClient` implementaions or
statuses simpler. Key changes include:

- Removing the logic that would enable and disable timers such as the
`reconnectionTimer` and `updateCapacityTimer` based on the status.
Instead, the timers are always running, but only take action if the
status matches a specified condition.

- Making all status change calls synchronous. When a particular status
involves an asynchronous component, a method that wraps that status
change with the asynchronous logic can be used instead (see next point).

- A new `setConnected` method that handles all steps related to marking
a client as connected, including updating its capacity.

- An inherited `init` method that handles generic steps around
initializing a swap client. It calls an abstract `initSpecific` method
that is implemented by each class extending `SwapClient` to contain
the client-specific logic for initialization.

- Making the raiden client optional within the swap client manager.
When disabled, the raiden client will not be initialized at all.