feat(swapclient): add Locked status
This adds a status to the swap clients for after a client has been
unlocked but before we have verified its connectivity. There can be
additional initialization steps after a client has been unlocked and
a possibility that it will encounter an error and shutdown, so we can
not assume that the client will be available and set the status to
`ConnectionVerified`. However, leaving the status as `WaitingUnlock`
can be misleading or lead to additional unlock attempt that will
fail. The new `Unlocked` status prevents further unlock attempts.

Closes #1411.