fix: allow unlock with disconnected lnd
This corrects unintentional behavior that prevented xud from unlocking
when any enabled instance of lnd is disconnected or unavailable when we
go to unlock it. This behavior is desired for the `create` and `restore`
calls - we want to ensure that we fully initialize and restore all
wallets. However, for unlock calls we can start up even if one lnd
instance is temporarily unavailable. When it comes back online it will
be automatically unlocked and activated to be prepared for swaps.