fix(lnd): exponential backoff for grpc client wait
This improves the logic behind waiting for an lnd grpc client to be
ready so that it is more robust. Occasionally the callback is never
triggered in cases where lnd is abruptly shut down and restarted
immediately, meaning we never attempt to reconnect to lnd because we are
never notified that the client is ready. To prevent this and any other
flakiness we set a max time to wait for the client to be ready and
attempt a grpc call regardless after this time. The time uses an
exponential backoff to prevent repeated calls to an offline server.

Closes #1415.