feat(cli): automatically determine rpc host/port
This enhances the cli to automatically attempt to find and read from
the xud configuration file to determine the port and host for the xud
gRPC service based on the configured network - which determines default
port numbers - or configured port if specified. This prevents the user
from having to specify the rpc port on `xucli` commands when using a
non-mainnet xud instance or a non-default gRPC port via the config file.

Closes #1451.