fix(lnd): relax unhandled openchannel timeout
This removes the hard timeout placed on each attempt to connect to a
peer's lnd node when attempting to open a channel with that peer. This
timeout was causing xud to crash because it threw an error that was not
being handled in the encapsulating catch block, which did not catch the
error because it was thrown from a callback.

Rather than attempt to correct the asynchronous exception handling,
this lifts the timeout limit set within xud and instead will wait on
each connect attempt until it succeeds, times out, or fails according to
the logic within lnd.

Fixes #1405.