fix: string message grpc error handling
This fixes an internal bug where passing a string to the `getGrpcError`
method - or possibly any object without `code`, `message`, or `name`
properties - would result in a gRPC exception of `TypeError: Incorrectly
typed arguments to startBatch` that is unhandled and crashes xud. This
ensures that the gRPC errors returned by xud are well-formed.