fix(cli): use localhost when rpc config 0.0.0.0
This fixes an issue where xucli could read an rpc host of `0.0.0.0` from
the config file, indicating that xud is listening on all interfaces,
and then attempt to connect to it using that ip address which fails.
Instead, if the xucli tool detects that xud is configured to listen on
all interfaces, it should default to attempting to connect on localhost.