fix: resolve alias for all known nodes
This fixes the logic used to resolve a node alias - when it is provided
via the rpc layer, to a node pub key. Previously, we only checked if
the alias matched connected peers and banned nodes, but not nodes we
know of but aren't actively connected to.

This also refactors the `resolveAlias` logic to use a map of known
aliases to node pub keys, such that we don't dynamically calculate the
list of known aliases each time we must resolve an alias. Instead,
aliases for all known nodes are determined up front.

Finally it refactors the nodelist to combine the `Map` of banned nodes
with that of unbanned nodes, as the distinction was unnecessary. Instead
a single map contains all known nodes.