fix(swaps): don't cancel invoice on swap timeout
This fixes a bug where we cancel the incoming invoice for swaps that
time out as the maker. Although we would put the timed out swap into
the SwapRecovery module, the swap cannot be recovered if the incoming
invoice has been canceled, and if the taker later claims our payment
it leaves us unable to claim the maker's payment.

The widely used `failDeal` method in Swaps was always canceling
incoming invoices which is what led to this unintended behavior.
Canceling invoices should only happen when we are certain that our
payment won't go through or when we haven't begun our payment yet.

Fixes #1574.