chore: print xud logs only for simulation tests
This outputs only the xud logs for the `test:sim:logs` script, ignoring
the lnd logs since they are too verbose and can result in too much text
in the travis log output when the script is used. This also prints the
filename at the top of each log file so they can be differentiated.