feat(swaps): use SwapRecovery for maker pay errors
This removes all code from the `Swaps` module that would cancel an
invoice as maker when encountering a `sendPayment` failure. Instead, we
put swaps that fail when sending a payment from maker to taker into
`SwapRecovery`, where incoming invoices are only canceled after
confirming that payment has failed permanently via a `lookupPayment`
call. This is to ensure that the maker never gives up the ability to
claim an incoming payment while the corresponding outgoing payment
can still be claimed by the taker.