feat(swaps): check payment immediately on recovery
This commit makes it so that any swap deals that fail as maker while
sending payment will have the payment checked immediately upon being
added to set of swaps tracked by the `SwapRecovery` module. This is
instead of waiting for the 30 second timer to periodically retry tracked
swaps

It also refactors the code around checking and rechecking swaps that
may have a pending payment involved, and enhances logging output.

Closes #1598.