fix(swaps): don't try to accept a failed deal
This ensures that we don't attempt to advance an accepted swap that we
have already failed, as can be the case when we fail a deal that timed
out during the acceptance stage. If the maker accepts our swap after
we already timed out and failed the deal, this situation can arise.

This can cause xud to crash due to an AssertionError from attempting to
change the phase of a swap that has already failed.

Fixes #1614.