refactor(swaps): check inactive currencies sooner
This moves the check to ensure that we are connected to the swap clients
for both sides of a swap before accepting a deal request to before we
create the swap deal object and begin the swap timer. It also makes
some minor refactoring of the surrounding code and variables to make
it more concise and readable.