fix(backup): use async fs calls
This fix attempts to address issues and timeouts related to the Backup
module writing new backups to disk synchronously. This can block the
thread and prevent xud from proceeding with and responding to other
matters requiring its attention. It appears to cause the swap acceptance
logic to occasionally exceed the 10 second limit because accepting a
swap involves several db writes which trigger the backup writes.