refactor(swaps): parallel swap accept packet send
This refactors the final steps of the routine to accept a swap deal to
make it so that we update the deal phase (and in turn, update the db)
in parallel with sending the swap accepted packet to our peer. This
can reduce the total time needed to accept a deal request and begin the
swap.