fix(swaps): cancel timer immediately on complete
This changes the order of events when a swap enters the `PaymentReceived`
phase and is considered complete. Currently, we mark the swap as
complete right away but we don't cancel the swap timer until after we've
updated the database to persist the swap. When the database write is
delayed at all, it presents the edge case risk where the swap timer
hits its limit while are writing to the database and after we've already
marked the swap as complete. This results in an assertion error as we
try to fail a swap due to a timeout when that swap is already complete.

Instead, we cancel the swap timer immediately after marking the swap
complete and only then persist the update to the database.

Fixes #1634.