fix(cli): streamorders handling xud not ready
This fixes the behavior of the `streamorders` cli command when xud is
not ready for the call. Previously, if xud was locked it would stream
an endless series of reconnecting messages and if xud was not running
on the specified host/port at all it would simply hang. Now the command
exits with an appropriate error message in such cases.