feat: penalize peers for delaying preimage release
This creates two new reputation events for penalizing peers that
release the preimage for a swap after it was supposed to have timed out
and failed. This indicates that a peer may be running malicious code
intended to exploit the free options problem. The more severe event
for an abnormally long delay (currently set to more than a minute past
the expected swap timeout) results in an instant ban.