feat(p2p): improved ban handling
This makes several improvements to the way xud behaves when a node is
banned.

1. Only emit the `node.ban` event a single time when a node is banned,
previously it was possible for this to fire multiple times when multiple
negative reputation events occur at roughly the same time, polluting the
logs and sending multiple disconnecting packets to the peer.

2. Log the name of the reputation event that triggered the ban.

3. Only send the recent reputation events to a peer when we ban them,
rather than JSON stringifying the entire ReputationEventInstance db
object which contains a lot of internal metadata that's irrelevent to
peers.